Game autowinner hook fix
